refactor diff logic into new stream class
